mlb-pipeline/requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
supabase>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
urllib3>=2.0.0
xgboost>=2.0.0
scikit-learn>=1.3.0
